+++ Source/Mosa.UnitTests/ForeachTests.cs
@@ -101,7 +101,7 @@ namespace Mosa.UnitTests
 		{
 			int[] a = new int[5];
 			for (int i = 0; i < 5; i++)
-				a[i] = (int)i;
+				a[i] = i;
 
 			int total = 0;
 
@@ -116,7 +116,7 @@ namespace Mosa.UnitTests
 		{
 			long[] a = new long[5];
 			for (int i = 0; i < 5; i++)
-				a[i] = (long)i;
+				a[i] = i;
 
 			long total = 0;
